Move text-editor from packages to app-development/features
Because it does not need a separate package.json file
