+++ src/studio/src/designer/frontend/app-development/features/text-editor/components/Row.tsx
