+++ src/studio/src/designer/frontend/app-development/features/text-editor/hooks/useEditArray.ts
